+++ opentile/formats/__init__.py
@@ -15,6 +15,7 @@
 """Module containing format specific implementations."""
 
 from opentile.formats.histech import HistechTiffTiler
+from opentile.formats.huron import HuronTiffTiler
 from opentile.formats.ndpi import NdpiTiler
 from opentile.formats.ome import OmeTiffTiler
 from opentile.formats.philips import PhilipsTiffTiler
@@ -26,6 +27,7 @@ __all__ = [
     "OmeTiffTiler",
     "NdpiTiler",
     "HistechTiffTiler",
+    "HuronTiffTiler",
     "PhilipsTiffTiler",
     "SvsTiler",
     "TrestleTiffTiler",