+++ tests/filepaths.py
@@ -26,6 +26,7 @@ ndpi_jpegxr_file_path = slide_folder.joinpath("ndpi/jpegxr1/jpegxr1.ndpi")
 histech_file_path = slide_folder.joinpath(
     "3dhistech_tiff/CMU-1/CMU-1_Default_Extended.tif"
 )
+huron_file_path = slide_folder.joinpath("huron/Huron-1/Huron-1.tif")
 ome_tiff_file_path = slide_folder.joinpath("ome_tiff/CMU-1/CMU-1.ome.tiff")
 trestle_file_path = slide_folder.joinpath("trestle_tiff/CMU-1/CMU-1.tif")
 ventana_file_path = slide_folder.joinpath("ventana_bif/OS-1/OS-1.bif")